apps/genai-t5-model/requirements.txt
chromadb-client==0.4.14.dev0
cmake==3.27.5
jsonpatch==1.33
jsonpointer==2.4
kserve==0.11.1
label-studio-sdk==0.0.32
label-studio-tools==0.0.3
langchain==0.0.302
langsmith==0.0.40
nltk==3.8.1
numpy==1.26.0
requests==2.31.0
safetensors==0.3.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
SQLAlchemy==2.0.21
tenacity==8.2.3
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers==4.33.2
triton==2.0.0
urllib3==2.0.5